Add MongoDB support and slightly change behavior

+++ aiogram_media_group/handler.py
@@ -27,6 +27,23 @@ elif AIOGRAM_VERSION == 2:
     from aiogram.dispatcher import FSMContext
     from aiogram.dispatcher.storage import BaseStorage as AiogramBaseStorage
     from aiogram.contrib.fsm_storage.memory import MemoryStorage as AiogramMemoryStorage
+
+    MONGO_INSTALLED = False
+    REDIS_INSTALLED = False
+
+    try:
+        from motor import motor_asyncio
+
+        from aiogram.contrib.fsm_storage.mongo import MongoStorage as AiogramMongoStorage
+
+        from .storages.mongo import MongoStorage
+    except ModuleNotFoundError:
+        # ignore if motor is not installed
+        pass
+    else:
+        MONGO_INSTALLED = True
+    
+
     try:
         import aioredis
 
@@ -39,33 +56,74 @@ elif AIOGRAM_VERSION == 2:
     except ModuleNotFoundError:
         # ignore if aioredis is not installed
         pass
+    else:
+        REDIS_INSTALLED = True
 
 
     async def _wrap_storage(storage: AiogramBaseStorage, prefix, ttl):
         storage_type = type(storage)
+        
         if storage_type is AiogramMemoryStorage:
             return MemoryStorage(data=storage.data, prefix=prefix)
-        elif storage_type is AiogramRedisStorage:
-            connection: aioredis.RedisConnection = await storage.redis()
-            return RedisStorage(connection=connection, prefix=prefix, ttl=ttl)
-        elif storage_type is AiogramRedis2Storage:
-            redis: aioredis.Redis = await storage.redis()
-            return RedisStorage(connection=redis.connection, prefix=prefix, ttl=ttl)
+        
+        elif MONGO_INSTALLED:
+            if storage_type is AiogramMongoStorage:
+                mongo: motor_asyncio.AsyncIOMotorDatabase = await storage.get_db()
+                return await MongoStorage.init(db=mongo)
+        
+        elif REDIS_INSTALLED:
+            if storage_type is AiogramRedisStorage:
+                connection: aioredis.RedisConnection = await storage.redis()
+                return RedisStorage(connection=connection, prefix=prefix, ttl=ttl)
+            elif storage_type is AiogramRedis2Storage:
+                redis: aioredis.Redis = await storage.redis()
+                return RedisStorage(connection=redis.connection, prefix=prefix, ttl=ttl)
+        
         else:
             raise ValueError(f"{storage_type} is unsupported storage")
+
+
+# to avoid errors and multiple calls to the same function
+handled_media_groups = []
+EXECUTING_CALLBACK = False
 
 
 async def _on_media_group_received(
     media_group_id: str,
     storage: BaseStorage,
     callback,
+    event_loop,
+    unhandle_timeout: float,
     *args,
     **kwargs,
 ):
-    messages = await storage.get_media_group_messages(media_group_id)
-    await storage.delete_media_group(media_group_id)
+    global EXECUTING_CALLBACK
 
-    return await callback(messages, *args, **kwargs)
+    while EXECUTING_CALLBACK:
+        await asyncio.sleep(0)
+    else:
+        EXECUTING_CALLBACK = True
+    
+    if not media_group_id in handled_media_groups:
+        handled_media_groups.append(media_group_id)
+        EXECUTING_CALLBACK = False
+
+        messages = await storage.get_media_group_messages(media_group_id)
+        await storage.delete_media_group(media_group_id)
+
+        event_loop.call_later(
+            unhandle_timeout,
+            asyncio.create_task,
+            unhandle_media_group(media_group_id),
+        )
+
+        return await callback(messages, *args, **kwargs)
+        
+    EXECUTING_CALLBACK = False
+
+
+async def unhandle_media_group(media_group_id: str):
+    handled_media_groups.remove(media_group_id)
 
 
 def media_group_handler(
@@ -102,7 +160,7 @@ def media_group_handler(
                     receive_timeout,
                     asyncio.create_task,
                     _on_media_group_received(
-                        message.media_group_id, storage, handler, *args, **kwargs
+                        message.media_group_id, storage, handler, event_loop, 2.0, *args, **kwargs
                     ),
                 )
